+++ src/App.tsx
@@ -293,13 +293,15 @@ class App extends React.Component <{
 
 	getCoursesBySelectedShifts(): Course[] {
 		const finalCourses = [...this.state.selectedCourses.courses]
-		// let courses: Record<string, Course> = {}
+
 		this.state.selectedShifts.forEach( (s) => {
 			// FIXME: Includes? hmmmm
 			// finalCourses = Comparables.addToSet(finalCourses, s.course) as Record<string, Course>
 			if (!finalCourses.includes(s.course)) {
 				finalCourses.push(s.course)
 			}
+			// Update course shift types (if selected or not) when the course is added and there was already shifts selected
+			s.course.addSelectedShift(s)
 		})
 		return finalCourses.sort(Course.compare)
 	}